Fixed black times & thumbler dialog for turns per addition
Black initial and additional times chosen in the time picker now
delivered to the new game dialog. Also the texts for black are no longer
updated when equal times is chosen, as it is not the text that is used
to get the values. (Planning to hide black items when equal times, so no
replacement.)